Private property accessor should trigger clone
When accessing a property's value storage directly by using
the underscored version of the property name, we still need
to clone the default value to prevent it being reused. The
best way to do this is to make the internal accessor a property
itself and move value storage to a double-underscored version of
the property name.