Maintain collection order for ListView insertion
ListViewItems should be inserted into the DOM in the same order
they are stored in the collection. This adds an @ordered property
to ListView that does just that when it's set to true. Setting @ordered 
to false maintains the current behavior of doing a simple append to
the bottom of the list of elements.