Avoid hijacking clicks in non-selectable ListView
A link embedded in a ListViewItem will not be clickable by default as
the click intention is assumed to be to select that ListViewItem and the
event is prevented from bubbling further. When the ListView property
allowsSelection is set to "no", we should avoid hijacking that click and
allow it to pass through to any element contained within the
ListViewItem.